Started breaking apart `functions` index file
